# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /workplace/:
    put:
      tags:
        - Рабочее место
      summary: Изменить указанное место
      description: Все поля обязательны
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkPlaceUpdateDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    post:
      tags:
        - Рабочее место
      summary: Создание рабочего места
      description: Все поля кроме имени обязательны, имя по дефолту "№ "{newId}""
      operationId: createWorkPlace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkPlaceCreateDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                type: integer
                format: int64
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
  /workPlaceType/:
    put:
      tags:
        - Тип места
      summary: Изменить тип места
      operationId: update_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkPlaceTypeDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    post:
      tags:
        - Тип места
      summary: Добавление нового типа места
      operationId: createType
      parameters:
        - name: typeName
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /workplace/{id}:
    get:
      tags:
        - Рабочее место
      summary: Получить указанное места
      operationId: getWorkPlaceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkPlaceGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    delete:
      tags:
        - Рабочее место
      summary: Удалить указанное место
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /workplace/all:
    get:
      tags:
        - Рабочее место
      summary: Все рабочие места
      description: 1 <= size <= 20 (default 20)
      operationId: getPage
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageWorkPlaceGetDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /workPlaceType/{id}:
    get:
      tags:
        - Тип места
      summary: Получить тип места по id
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkPlaceTypeDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
    delete:
      tags:
        - Тип места
      summary: Удалить тип места
      operationId: delete_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
  /workPlaceType/name:
    get:
      tags:
        - Тип места
      summary: Получить тип места по названию
      operationId: getByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageWorkPlaceTypeDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
  /workPlaceType/all:
    get:
      tags:
        - Тип места
      summary: Получить все типы мест
      operationId: getAll
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: >-
            Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported.
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageWorkPlaceTypeDto'
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
  /favPlace/{employeeId}/place/{placeId}:
    delete:
      tags:
        - Избранное место
      summary: Убрать место из избранного по его id и id сотрудника
      operationId: delete_7
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: placeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: >-
            Ошибки валидации. Два варианта возвращаемого JSON объекта (смотреть
            Examples)
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ExceptionDto'
                  - $ref: '#/components/schemas/ExceptionsDto'
              examples:
                Вариант с одной ошибкой:
                  description: Вариант с одной ошибкой
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    message: Что-либо не прошло валидацию.
                Вариант со списком ошибок:
                  description: Вариант со списком ошибок
                  value:
                    status: BAD_REQUEST
                    timestamp: 08-08-2008 08:08:08
                    errors:
                      field: Поле field не прошло валидацию.
                      name: Поле name не прошло валидацию.
        '404':
          description: Объект не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
        '409':
          description: Объект уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionDto'
...
components:
  schemas:
    ExceptionDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    ExceptionsDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          format: date-time
        errors:
          type: object
          additionalProperties:
            type: string
    WorkPlaceUpdateDto:
      required:
        - capacity
        - floorId
        - id
        - name
        - typeId
      type: object
      properties:
        id:
          minimum: 1
          type: integer
          format: int64
        name:
          type: string
        typeId:
          minimum: 1
          type: integer
          format: int64
        floorId:
          minimum: 1
          type: integer
          format: int64
        capacity:
          minimum: 1
          type: integer
          format: int32
    WorkPlaceTypeDto:
      required:
        - id
        - name
      type: object
      properties:
        id:
          minimum: 1
          type: integer
          format: int64
        name:
          type: string
...
    WorkPlaceCreateDto:
      required:
        - capacity
        - floorId
        - typeId
      type: object
      properties:
        placeName:
          type: string
        typeId:
          minimum: 1
          type: integer
          format: int64
        floorId:
          minimum: 1
          type: integer
          format: int64
        capacity:
          minimum: 1
          type: integer
          format: int32
...
    WorkPlaceGetDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        placeName:
          type: string
        typeName:
          type: string
        floorId:
          type: integer
          format: int64
        capacity:
          type: integer
          format: int32
...
    PageWorkPlaceGetDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/WorkPlaceGetDto'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...